Revert "test: auth web"
